Fix CommandContainer badly handling abstract classes
Before this commit, if an abstract class was extending LinkableCommand or any of its children, the abstract getCalls function would crash as it might not have been initialized yet.

We now silently fail those cases, as we don't need to do anything particuliar for those instances.

The links object is now a ReadonlyMap, which helps in making this look more object oriented, since this is basically what this is.

Also added few test cases for this particular behavior.

Also removed the getCalls and action functions of the AbstractBotCommand, as we want them to be defined by the subclasses themselves.