Tweak parameters to use explicit Type<T> declaration
The CommandContainer now defines the generic type Type<T> which takes a class and force the given type to be a function that can be initialized of that type.

The getLinkableCommandClass should now also more accurately distinguish from actual LinkableCommand classes and other objects.

The createFromDir function now also allows to provide additional classes / folders to load via a new optional parameter.