Add frameworke managers for common Bot creation steps
Add : BotManager, to handle creating a bot, a default listener if required, and common methods to start / stop the bot.

Add : DiscordEventListener, to handle events generated by the Discord client, with methods to easily add listeners or change the default onMessage function. This also creates a default CommandRouter if no custom command router is specified.

Add : CommandRouter, to route the message to the correct command, and initializes the command with various bits of data, such as the request, the router itself, etc.

Tweak : LinkableCommand now defines the options as an OptionDef type, which uses the vcommand-parser package to deal with requests / options.

Tweak : Generic types declarations are now in their own file for utils purposes.