Add CommandRouterOptions, which allows for setting custom prefixes
These custom prefixes can be either a permanent affair or a function that is triggered on each message received with the message given as context (useful for creating keys for searching through settings, for example).

AbstractBotCommand now has more context about the available variables, such as the message of the router and the client inside it.