Add callCommand function
This callCommand function allows commands to call other commands, allowing for some interesting chaining.

If the command requested doesn't exists in the CommandContainer instance, an error is passed to the `then` function of the Promise. IT DOES NOT MAKE THE COMMAND FAIL. This is done to allow for silently failing command calls, or in other words, allowing to manually handle how a missing command is handled.